First of many updates to pathfinding Pathfinding has recieved an upgrade here. We now check if they can reach their destination before making them walk over, this prevents them from dogpiling on the surface when you're minding your own business in an unreachable cave.
Next update will be an overhaul to this system now that i know you can mess with the pathfinding. Instead of using the target system to get them to move, we'll use the the pathfind system and move them towards the alert location. The dream is that this mimics how a zombie might actually react to sound.